Move notifications into separate file with new name
